+++ src/command_modules/azure-cli-acs/HISTORY.rst
@@ -5,7 +5,9 @@ Release History
 
 2.0.26
 ++++++
+* `aks create` defaults to letting the server choose the version of Kubernetes
 * `aks create` VM node size default changed from "Standard_D1_v2" to "Standard_DS1_v2"
+* improve reliability when locating the dashboard pod for `az aks browse`
 
 2.0.25
 ++++++