+++ src/command_modules/azure-cli-acs/HISTORY.rst
@@ -12,6 +12,7 @@ Release History
 * `aks create` VM node size default changed from "Standard_D1_v2" to "Standard_DS1_v2"
 * improve reliability when locating the dashboard pod for `az aks browse`
 * `aks get-credentials` handles UnicodeDecodeError when loading Kubernetes configuration files
+* add a message to `az aks install-cli` to help get `kubectl.exe` in the search PATH
 
 2.0.25
 ++++++